Refactor: extract types to module
